Refine tree truncation
